pyproject.toml
[tool.poetry]
name = "tesla-project"
version = "0.1.0"
description = ""
authors = ["Ihor Voitiuk"]
readme = "README.md"
packages = [{include = "english_school"}]

[tool.poetry.dependencies]
python = "^3.9"
django = "4.1"
django-environ = "^0.10.0"
pillow = "^10.0.0"
djangorestframework = "^3.14.0"
markdown = "^3.4.3"
psycopg2-binary = "^2.9.6"
django-jet-reboot = "^1.3.3"
six = "^1.16.0"
feedparser = "^6.0.10"
cloudinary = "^1.33.0"
django-cloudinary-storage = "^0.3.0"
gunicorn = "^20.1.0"
whitenoise = "6.3.0"
django-cors-headers = "^4.2.0"
asgiref = "^3.7.2"
django-parler = "^2.3"
django-tinymce = "^3.6.1"
google-api-python-client = "^2.101.0"
oauth2client = "^4.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
